# apps/backend/swagger.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"Успешная регистрация","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Неккоректные данные"}},"summary":"Регистрация пользователя","tags":["Auth","Аутентификация"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"Успешная авторизация","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Неккоректные данные"}},"summary":"Авторизация пользователя","tags":["Auth","Аутентификация"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"201":{"description":""}},"summary":"Обновление access_token по refreshToken cookie (rotation)","tags":["Auth","Аутентификация"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"summary":"Logout (удаляет refresh token)","tags":["Auth","Аутентификация"]}},"/auth/confirm-email":{"get":{"operationId":"AuthController_confirmEmail","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Подтверждение email","tags":["Auth","Аутентификация"]}},"/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":""}},"summary":"Текущий пользователь (по access token)","tags":["Auth","Аутентификация"]}},"/auth/resend-confirmation":{"post":{"operationId":"AuthController_resendConfirmation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendConfirmationDto"}}}},"responses":{"201":{"description":""}},"summary":"Повторная отправка письма подтверждения","tags":["Auth","Аутентификация"]}}},"info":{"title":"Родная Ярмарка","description":"Документация REST API","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"you@mail.ru","description":"Почта пользователя"},"password":{"type":"string","example":"qwerty123","description":"Пароль пользователя"},"name":{"type":"string","example":"Иван","description":"Имя пользователя"}},"required":["email","password","name"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"you@mail.ru","description":"Почта пользователя"},"password":{"type":"string","example":"qwerty123","description":"Пароль пользователя"}},"required":["email","password"]},"ResendConfirmationDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]}}}}